move core GUI components into gui submodule
