+++ game/src/app.rs
@@ -6,7 +6,7 @@ use crate::{game::GameState, gui::*, TextureMap};
 use conrod_core::{color::Colorable, widget, widget::Widget, Borderable};
 
 use conrod_core::input::Key;
-use learning_conrod_core::{cache_queued_glyphs, Application, RenderContext, GUI};
+use learning_conrod_core::gui::{cache_queued_glyphs, Application, RenderContext, GUI};
 use piston::event_loop::Events;
 use piston::input::{Event, Input, RenderArgs, UpdateArgs};
 use piston_window::PistonWindow;
